# docs/components/backend/identity-resolution/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateVisibilityRequest": {
        "description": "Body of `POST /v1/visibility` — grant a viewer visibility over a target\n(or the whole tree when `viewed_person_id` is omitted).",
        "properties": {
          "reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "valid_from": {
            "description": "Optional grant start; defaults to now when omitted. Accepts RFC-3339\n(`Z`/offset), zone-less, or date-only, normalised to naive-UTC.",
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "viewed_person_id": {
            "format": "uuid",
            "type": [
              "string",
              "null"
            ]
          },
          "viewer_person_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "viewer_person_id"
        ],
        "type": "object"
      },
...
      "Problem": {
        "description": "RFC 9457 problem+json. `context` varies by error category.",
        "properties": {
          "context": {
            "type": "object"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "status": {
            "format": "int32",
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "trace_id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "context"
        ],
        "type": "object"
      },
...
      "RevokeReasonRequest": {
        "description": "Optional `DELETE` body carrying a revoke reason.",
        "properties": {
          "reason": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
      "VisibilityListResponse": {
        "description": "List wrapper.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/VisibilityResponse"
            },
            "type": "array"
          },
          "next_cursor": {
            "description": "Wire parity with the .NET `ListResponse`: the cursor is declared\nbut pagination is not implemented — always `null` (both\nimplementations return every row; consumers already tolerate it).",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
      "VisibilityResponse": {
        "description": "One visibility grant.",
        "properties": {
          "author_person_id": {
            "format": "uuid",
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "insight_tenant_id": {
            "format": "uuid",
            "type": "string"
          },
          "reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "valid_from": {
            "type": "string"
          },
          "valid_to": {
            "type": [
              "string",
              "null"
            ]
          },
          "viewed_person_id": {
            "format": "uuid",
            "type": [
              "string",
              "null"
            ]
          },
          "viewer_person_id": {
            "format": "uuid",
            "type": "string"
          },
          "visibility_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "visibility_id",
          "insight_tenant_id",
          "viewer_person_id",
          "valid_from",
          "author_person_id",
          "created_at"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Person identity for the product: profile resolution, the operator correction surface over account-to-person bindings, org-chart reads, roles and visibility, plus the persons-seed and persons-sync operation journals. The API Gateway mounts this service at /api/identity.",
    "title": "Insight Identity Resolution API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/visibility": {
      "get": {
        "operationId": "identity_resolution.visibility.list",
        "parameters": [
          {
            "description": "Only grants held by this viewer",
            "in": "query",
            "name": "viewer",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Only grants over this person",
            "in": "query",
            "name": "viewed",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Only grants still in force",
            "in": "query",
            "name": "active",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "description": "Cap on returned grants (1..=500, default 50)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisibilityListResponse"
                }
              }
            },
            "description": "Grants"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List visibility grants (admin)"
      },
      "post": {
        "operationId": "identity_resolution.visibility.create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVisibilityRequest"
              }
            }
          },
          "description": "Grant to create",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisibilityResponse"
                }
              }
            },
            "description": "Created grant"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a visibility grant (admin)"
      }
    },
    "/v1/visibility/{id}": {
      "delete": {
        "operationId": "identity_resolution.visibility.delete",
        "parameters": [
          {
            "description": "Visibility grant id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeReasonRequest"
              }
            }
          },
          "description": "Why the grant is revoked"
        },
        "responses": {
          "204": {
            "description": "Grant revoked"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Revoke a visibility grant (admin)"
      }
    },
...
  }
}